allow Markdown files preview in code browser
