+++ components/Package/MarkdownContentBox/index.tsx
@@ -1,42 +1,19 @@
-import { Md } from '@m2d/react-markdown/client';
-import { capitalize } from 'es-toolkit/string';
 import { useRouter } from 'next/router';
-import { Children, isValidElement, useEffect, useMemo, useState } from 'react';
+import { useEffect, useMemo, useState } from 'react';
 import { View } from 'react-native';
-import { type Theme } from 'react-shiki';
-import rehypeRaw from 'rehype-raw';
-import remarkEmoji from 'remark-emoji';
-import remarkGfm from 'remark-gfm';
 import useSWR from 'swr';
 
-import { A, P } from '~/common/styleguide';
-import {
-  CautionBlockquote,
-  CCFile,
-  ChangelogFile,
-  Check,
-  ContributingFile,
-  ImportantBlockquote,
-  NoteBlockquote,
-  ReadmeFile,
-  TipBlockquote,
-  WarningBlockquote,
-} from '~/components/Icons';
+import { P } from '~/common/styleguide';
+import { CCFile, ChangelogFile, ContributingFile, ReadmeFile } from '~/components/Icons';
 import CopyButton from '~/components/Package/CopyButton';
 import ThreeDotsLoader from '~/components/Package/ThreeDotsLoader';
-import rndDark from '~/styles/shiki/rnd-dark.json';
-import rndLight from '~/styles/shiki/rnd-light.json';
 import { type LibraryType, type MarkdownTab, type MarkdownTabsType } from '~/types';
 import { TimeRange } from '~/util/datetime';
-import { extractAndStripBlockquoteType } from '~/util/extractAndStripBlockquoteType';
-import { getReadmeAssetURL } from '~/util/getReadmeAssetUrl';
 import { parseGitHubUrl } from '~/util/parseGitHubUrl';
 import tw from '~/util/tailwind';
 
-import MarkdownCodeBlock from './MarkdownCodeBlock';
 import MarkdownContentTab from './MarkdownContentTab';
-import MarkdownHeading from './MarkdownHeading';
-import { MarkdownVideoPlayer } from './MarkdownVideoPlayer';
+import MarkdownRenderer from './MarkdownRenderer';
 import { DEFAULT_MARKDOWN_TAB, MARKDOWN_CONTENT_QUERY_PARAM, parseMarkdownTab } from './utils';
 
 type Props = {
@@ -196,165 +173,7 @@ export default function MarkdownContentBox({ packageName, library, loader = fals
             <P style={tw`text-center`}>{readmeFallbackContent}</P>
           </View>
         ) : (
-          <Md
-            id="markdownContentContainer"
-            components={{
-              h1: ({ children, node }: any) => (
-                <MarkdownHeading tagName={node.tagName}>{children}</MarkdownHeading>
-              ),
-              h2: ({ children, node }: any) => (
-                <MarkdownHeading tagName={node.tagName}>{children}</MarkdownHeading>
-              ),
-              h3: ({ children, node }: any) => (
-                <MarkdownHeading tagName={node.tagName}>{children}</MarkdownHeading>
-              ),
-              h4: ({ children, node }: any) => (
-                <MarkdownHeading tagName={node.tagName}>{children}</MarkdownHeading>
-              ),
-              h5: ({ children, node }: any) => (
-                <MarkdownHeading tagName={node.tagName}>{children}</MarkdownHeading>
-              ),
-              h6: ({ children, node }: any) => (
-                <MarkdownHeading tagName={node.tagName}>{children}</MarkdownHeading>
-              ),
-              p: ({ children, node, ...props }: any) => {
-                const childrenCount = Children.count(children);
-                if (childrenCount === 1) {
-                  const element = Children.toArray(children).at(0);
-                  if (
-                    isValidElement<{ href?: string }>(element) &&
-                    isGitHubVideoAssetLink(element.props.href)
-                  ) {
-                    return children;
-                  }
-                }
-
-                return <p {...props}>{children}</p>;
-              },
-              br: () => null,
-              hr: () => null,
-              a: (props: any) => {
-                if (props.href && !props.href.startsWith('//')) {
-                  if (!props.href.startsWith('http')) {
-                    return (
-                      <A
-                        {...props}
-                        href={`${repoUrl}/blob/HEAD/${props.href.startsWith('/') ? props.href.slice(1) : props.href}`}
-                      />
-                    );
-                  } else if (isGitHubVideoAssetLink(props.href)) {
-                    return <MarkdownVideoPlayer src={props.href} />;
-                  }
-                  return <A {...props} />;
-                }
-                return <span>{props.children}</span>;
-              },
-              video: (props: any) => {
-                if (props.src) {
-                  return <MarkdownVideoPlayer src={props.src} />;
-                }
-                return null;
-              },
-              table: ({ children }) => {
-                return (
-                  // @ts-expect-error dataSet is a valid RNW prop
-                  <View dataSet={{ tableWrapper: true }}>
-                    <table>{children}</table>
-                  </View>
-                );
-              },
-              img: ({ src, alt, width, height }: any) => (
-                <img
-                  src={getReadmeAssetURL(src, repoUrl)}
-                  onError={(error: any) => {
-                    const fallbackUrl = getReadmeAssetURL(src, repoUrl, 'master');
-                    const target = error.currentTarget;
-
-                    if (target.src !== fallbackUrl) {
-                      target.onerror = null;
-                      target.src = fallbackUrl;
-                    } else {
-                      target.style.display = 'none';
-                    }
-                  }}
-                  alt={alt ?? ''}
-                  width={width}
-                  height="auto"
-                  style={{
-                    maxHeight: height,
-                  }}
-                />
-              ),
-              source: ({ srcSet, ...rest }: any) => (
-                <source
-                  srcSet={srcSet ? getReadmeAssetURL(srcSet, repoUrl) : undefined}
-                  {...rest}
-                />
-              ),
-              pre: ({ children }: any) => {
-                const langClass = children?.props?.className;
-                if (langClass) {
-                  return (
-                    <MarkdownCodeBlock
-                      code={children.props.children}
-                      theme={(tw.prefixMatch('dark') ? rndDark : rndLight) as Theme}
-                      lang={langClass ? (langClass.split('-')[1] ?? 'sh').toLowerCase() : 'sh'}
-                    />
-                  );
-                }
-                return (
-                  <div style={tw`relative my-2`} className="readme-code-block">
-                    <pre className="shiki">{children}</pre>
-                  </div>
-                );
-              },
-              blockquote: ({ children }: any) => {
-                const blockquoteType = extractAndStripBlockquoteType(children);
-                return (
-                  <blockquote
-                    className={blockquoteType.type}
-                    style={{
-                      ...tw`text-secondary`,
-                      ...(blockquoteType.type
-                        ? {}
-                        : tw`border-palette-gray4 dark:border-secondary`),
-                    }}>
-                    {blockquoteType.type && (
-                      <strong className="blockquote-title" style={tw`flex items-center gap-1.5`}>
-                        {getBlockquoteIcon(blockquoteType.type)}
-                        {capitalize(blockquoteType.type)}
-                      </strong>
-                    )}
-                    {blockquoteType.children}
-                  </blockquote>
-                );
-              },
-              details: ({ children }: any) => {
-                return (
-                  <details
-                    style={tw`mt-3 rounded-xl border border-palette-gray2 pb-3 pr-4 pt-1 dark:border-default`}>
-                    {children}
-                  </details>
-                );
-              },
-              input: ({ type, checked, ...rest }: any) => {
-                if (type === 'checkbox') {
-                  const isChecked = Boolean(checked);
-                  return (
-                    <div
-                      className="checkbox"
-                      style={tw`m-0 box-border size-4 items-center justify-center rounded border border-solid border-palette-gray3 bg-palette-gray2 dark:border-powder dark:bg-palette-gray6`}>
-                      {isChecked && <Check style={tw`size-3 text-success`} />}
-                    </div>
-                  );
-                }
-                return <input type={type} {...rest} />;
-              },
-            }}
-            rehypePlugins={[rehypeRaw]}
-            remarkPlugins={[remarkGfm, remarkEmoji]}>
-            {data ?? undefined}
-          </Md>
+          <MarkdownRenderer data={data} repoUrl={repoUrl} />
         )}
       </View>
     </View>
@@ -380,28 +199,4 @@ function getReadmeFallbackContent(
 function getTabContentUrl(library: LibraryType, fileName: string) {
   const { packagePath, branchName } = parseGitHubUrl(library.githubUrl);
   return `${library.github.urls.repo?.replace('github.com/', 'raw.githubusercontent.com/')}/${branchName ?? 'HEAD'}/${packagePath !== '.' ? `${packagePath}/` : ''}${fileName}`;
-}
-
-function getBlockquoteIcon(type: string) {
-  switch (type) {
-    case 'note':
-      return <NoteBlockquote style={tw`-ml-0.5 size-4`} />;
-    case 'tip':
-      return <TipBlockquote style={tw`-ml-0.5 size-4`} />;
-    case 'warning':
-      return <WarningBlockquote style={tw`-ml-0.5 size-4`} />;
-    case 'caution':
-      return <CautionBlockquote style={tw`-ml-0.5 size-4`} />;
-    case 'important':
-      return <ImportantBlockquote style={tw`-ml-0.5 size-4`} />;
-    default:
-      return null;
-  }
-}
-
-function isGitHubVideoAssetLink(link?: string) {
-  if (!link) {
-    return false;
-  }
-  return link.startsWith('https://github.com/user-attachments/assets/');
 }
