+++ styles/styles.css
@@ -794,6 +794,19 @@ select {
     scrollbar-gutter: stable;
   }
 
+  #markdownContentWrapper {
+    scrollbar-color: var(--icon) var(--background);
+    scrollbar-gutter: stable;
+
+    code {
+      font-size: 83%;
+    }
+
+    pre code {
+      font-size: 110%;
+    }
+  }
+
   pre {
     margin-block: 0;
     padding: 16px;
